fix: instrument formatting, sync wiring, and rebase cleanup
- InstrumentAmount.formatted now instrument-kind-aware: fiat uses
  .currency(code:), stocks/crypto use "quantity symbol" format
- InstrumentAmountView updated to match
- Added precondition on +/- operators to catch instrument mismatches
- Extracted FiatConversionService and FullConversionService from
  Domain/Services/ to Shared/ (fixes dependency inversion)
- Wire onInstrumentChanged and leg record sync notifications for
  CKSyncEngine to sync InstrumentRecord and TransactionLegRecord
- Apply SyncErrorRecovery fix: re-queue on serverRecordChanged
  when serverRecord is nil instead of silently dropping
- Fix ProfileIndexSyncEngine buildCKRecord to use fresh record
  fields with cached system fields
- Remove cachedBalance invalidation code (balances computed from
  legs in multi-instrument model)
- Make SystemFieldsCacheable class-constrained (AnyObject)
- Migrate tests from old MonetaryAmount/Transaction API
- Remove obsolete cachedBalance sync tests
- Center crypto preferences "No Tokens" empty state
- Clean up BUGS.md (all items fixed or moved to main)
